fix linh tinh

+++ Cinema/Models/DataInfo/PageListComboDTO.cs
@@ -7,7 +7,7 @@ namespace CinemaAPI.Models.DataInfo
     {
         
         public string ComboName { get; set; } 
-        public int? Price { get; set; }
+        public double? Price { get; set; }
         public DateTime TimeCreated { get; set; }
         public sbyte Status { get; set; }
     }
@@ -17,7 +17,7 @@ namespace CinemaAPI.Models.DataInfo
         public string ComboName { get; set; }
         public string ComboItems { get; set; }
         public string ImageUrl { get; set; }
-        public int? Price { get; set; }
+        public double? Price { get; set; }
         public sbyte Status { get; set; }
     }
    public class ComboForBill : BaseDTO
